@@ -18,7 +18,8 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+ # sdk: ">=2.7.0 <3.0.0"
  # flutter: ">=2.0.0"
 
 dependencies:
@@ -48,7 +49,7 @@ dependencies:
   move_to_background: ^1.0.2
   shared_preferences: ^2.0.7
   #workmanager: ^0.4.0
-  flutter_local_notifications: ^8.1.1+1
+  #flutter_local_notifications: ^8.1.1+1
   map_markers: ^0.1.0
   flutter_background_service: ^0.1.5
   #services: ^0.0.1
@@ -61,7 +62,7 @@ dependencies:
   pedantic: ^1.11.1
   speech_bubble: ^0.0.9
   splashscreen: ^1.3.5
-  animated_dialog_box: ^0.0.2
+ # animated_dialog_box: ^0.0.2
   cool_alert: ^1.1.0
   flutter_holo_date_picker: ^0.3.4+1
   flutter_country_picker: ^0.1.6
@@ -70,11 +71,18 @@ dependencies:
   # firebase_messaging: ^10.0.1
   # firebase_core: ^1.6.0  
   dots_indicator: ^2.0.0
-  country_list_pick:
+ # country_list_pick:
  # country_codes: ^2.0.1
   country_code: ^1.0.0
   geocoder: ^0.2.1
   csc_picker: ^0.2.6
+  firebase_core: "^1.10.0"
+  firebase_messaging: ^11.2.0
+  flutter_local_notifications: ^9.1.4
+  #http: ^0.13.0
+  firebase_auth: ^3.3.0
+  firebase_analytics: ^8.3.4
+  cloud_firestore: ^3.1.1
 
 dev_dependencies:
   flutter_test:
