@@ -5,11 +5,11 @@ import 'globals.dart';
 import 'mobile.dart';
 
 class ApiMobile {
-  static Future<List<Mobile>> getMobiles() async {
+  static Future<List<Mobile>?> getMobiles() async {
     print("in apiMobile, getMobiles");
     try {
       final SharedPreferences prefs = await SharedPreferences.getInstance();
-      String username = prefs.getString('username');
+      String? username = prefs.getString('username');
       final res = await http
           .post(Uri.parse(getMobilesUrl), body: {"username": username});
       final body = json.decode(res.body);
